restructure third even setup
